create router constructor that accepts client
