--- webui/options.go
@@ -1,11 +1,0 @@
-package webui
-
-type opts func(*Server) *Server
-
-func WithPrefix(prefix string) opts {
-	return func(server *Server) *Server {
-		server.pathPrefix = prefix
-
-		return server
-	}
-}
