+++ webui/router.go
@@ -4,36 +4,39 @@ import (
 	"html/template"
 
 	"github.com/gocraft/web"
+	"github.com/gojek/work"
 )
 
-func (s *Server) setupRouter() {
-	s.router.Middleware(func(c *context, rw web.ResponseWriter, r *web.Request, next web.NextMiddlewareFunc) {
-		c.Server = s
-		next(rw, r)
-	})
-	s.router.Middleware(func(rw web.ResponseWriter, r *web.Request, next web.NextMiddlewareFunc) {
+type RouterOptions struct {
+	PathPrefix string
+}
+
+func NewRouter(client *work.Client, opts RouterOptions) *web.Router {
+	ctx := context{client: client}
+	router := web.New(ctx)
+
+	router.Middleware(func(rw web.ResponseWriter, r *web.Request, next web.NextMiddlewareFunc) {
 		rw.Header().Set("Content-Type", "application/json; charset=utf-8")
 		next(rw, r)
 	})
 
-	subRouter := s.router.Subrouter(context{}, s.pathPrefix)
-
-	subRouter.Get("/ping", (*context).ping)
-	subRouter.Get("/queues", (*context).queues)
-	subRouter.Get("/worker_pools", (*context).workerPools)
-	subRouter.Get("/busy_workers", (*context).busyWorkers)
-	subRouter.Get("/retry_jobs", (*context).retryJobs)
-	subRouter.Get("/scheduled_jobs", (*context).scheduledJobs)
-	subRouter.Get("/dead_jobs", (*context).deadJobs)
-	subRouter.Post("/delete_dead_job/:died_at:\\d.*/:job_id", (*context).deleteDeadJob)
-	subRouter.Post("/retry_dead_job/:died_at:\\d.*/:job_id", (*context).retryDeadJob)
-	subRouter.Post("/delete_all_dead_jobs", (*context).deleteAllDeadJobs)
-	subRouter.Post("/retry_all_dead_jobs", (*context).retryAllDeadJobs)
+	subRouter := router.Subrouter(ctx, opts.PathPrefix)
+	subRouter.Get("/ping", ctx.ping)
+	subRouter.Get("/queues", ctx.queues)
+	subRouter.Get("/worker_pools", ctx.workerPools)
+	subRouter.Get("/busy_workers", ctx.busyWorkers)
+	subRouter.Get("/retry_jobs", ctx.retryJobs)
+	subRouter.Get("/scheduled_jobs", ctx.scheduledJobs)
+	subRouter.Get("/dead_jobs", ctx.deadJobs)
+	subRouter.Post("/delete_dead_job/:died_at:\\d.*/:job_id", ctx.deleteDeadJob)
+	subRouter.Post("/retry_dead_job/:died_at:\\d.*/:job_id", ctx.retryDeadJob)
+	subRouter.Post("/delete_all_dead_jobs", ctx.deleteAllDeadJobs)
+	subRouter.Post("/retry_all_dead_jobs", ctx.retryAllDeadJobs)
 
 	//
 	// Build the HTML page:
 	//
-	assetRouter := subRouter.Subrouter(context{}, "")
+	assetRouter := subRouter.Subrouter(ctx, "")
 	assetRouter.Get("/", func(c *context, rw web.ResponseWriter, req *web.Request) {
 		rw.Header().Set("Content-Type", "text/html; charset=utf-8")
 
@@ -46,11 +49,8 @@ func (s *Server) setupRouter() {
 			return
 		}
 
-		data := struct {
-			PathPrefix string
-		}{
-			PathPrefix: c.pathPrefix,
-		}
+		// TODO: check if map works
+		data := struct{ PathPrefix string }{PathPrefix: opts.PathPrefix}
 
 		err = indexTemplate.Execute(rw, data)
 		if err != nil {
@@ -63,4 +63,6 @@ func (s *Server) setupRouter() {
 		rw.Header().Set("Content-Type", "application/javascript; charset=utf-8")
 		_, _ = rw.Write(mustAsset("work.js"))
 	})
+
+	return router
 }
