+++ webui/webui.go
@@ -5,52 +5,10 @@ import (
 	"fmt"
 	"net/http"
 	"strconv"
-	"sync"
-	"time"
 
-	"github.com/braintree/manners"
 	"github.com/gocraft/web"
-	"github.com/gojek/work"
 	"github.com/gojek/work/webui/internal/assets"
-	"github.com/gomodule/redigo/redis"
 )
-
-// Server implements an HTTP server which exposes a JSON API to view and manage gojek/work items.
-type Server struct {
-	namespace  string
-	pool       *redis.Pool
-	client     *work.Client
-	hostPort   string
-	server     *manners.GracefulServer
-	wg         sync.WaitGroup
-	router     *web.Router
-	pathPrefix string
-}
-
-type context struct {
-	*Server
-}
-
-// NewServer creates and returns a new server. The 'namespace' param is the redis namespace to use. The hostPort param is the address to bind on to expose the API.
-func NewServer(namespace string, pool *redis.Pool, hostPort string, serverOptions ...opts) *Server {
-	server := &Server{
-		namespace: namespace,
-		pool:      pool,
-		client:    work.NewClient(namespace, pool),
-		hostPort:  hostPort,
-		router:    web.New(context{}),
-	}
-
-	for _, serverOption := range serverOptions {
-		server = serverOption(server)
-	}
-
-	server.setupRouter()
-
-	server.server = manners.NewWithServer(&http.Server{Addr: hostPort, Handler: server.router})
-
-	return server
-}
 
 func mustAsset(name string) []byte {
 	b, err := assets.Asset(name)
@@ -58,153 +16,6 @@ func mustAsset(name string) []byte {
 		panic(err)
 	}
 	return b
-}
-
-// Start starts the server listening for requests on the hostPort specified in NewServer.
-func (w *Server) Start() {
-	w.wg.Add(1)
-	go func(w *Server) {
-		w.server.ListenAndServe()
-		w.wg.Done()
-	}(w)
-}
-
-// Stop stops the server and blocks until it has finished.
-func (w *Server) Stop() {
-	w.server.Close()
-	w.wg.Wait()
-}
-
-func (w *Server) Router() *web.Router {
-	return w.router
-}
-
-func (c *context) ping(rw web.ResponseWriter, r *web.Request) {
-	render(rw, map[string]string{"ping": "pong", "current_time": time.Now().Format(time.RFC3339)}, nil)
-}
-
-func (c *context) queues(rw web.ResponseWriter, r *web.Request) {
-	response, err := c.client.Queues()
-	render(rw, response, err)
-}
-
-func (c *context) workerPools(rw web.ResponseWriter, r *web.Request) {
-	response, err := c.client.WorkerPoolHeartbeats()
-	render(rw, response, err)
-}
-
-func (c *context) busyWorkers(rw web.ResponseWriter, r *web.Request) {
-	observations, err := c.client.WorkerObservations()
-	if err != nil {
-		renderError(rw, err)
-		return
-	}
-
-	var busyObservations []*work.WorkerObservation
-	for _, ob := range observations {
-		if ob.IsBusy {
-			busyObservations = append(busyObservations, ob)
-		}
-	}
-
-	render(rw, busyObservations, err)
-}
-
-func (c *context) retryJobs(rw web.ResponseWriter, r *web.Request) {
-	page, err := parsePage(r)
-	if err != nil {
-		renderError(rw, err)
-		return
-	}
-
-	jobs, count, err := c.client.RetryJobs(page)
-	if err != nil {
-		renderError(rw, err)
-		return
-	}
-
-	response := struct {
-		Count int64            `json:"count"`
-		Jobs  []*work.RetryJob `json:"jobs"`
-	}{Count: count, Jobs: jobs}
-
-	render(rw, response, err)
-}
-
-func (c *context) scheduledJobs(rw web.ResponseWriter, r *web.Request) {
-	page, err := parsePage(r)
-	if err != nil {
-		renderError(rw, err)
-		return
-	}
-
-	jobs, count, err := c.client.ScheduledJobs(page)
-	if err != nil {
-		renderError(rw, err)
-		return
-	}
-
-	response := struct {
-		Count int64                `json:"count"`
-		Jobs  []*work.ScheduledJob `json:"jobs"`
-	}{Count: count, Jobs: jobs}
-
-	render(rw, response, err)
-}
-
-func (c *context) deadJobs(rw web.ResponseWriter, r *web.Request) {
-	page, err := parsePage(r)
-	if err != nil {
-		renderError(rw, err)
-		return
-	}
-
-	jobs, count, err := c.client.DeadJobs(page)
-	if err != nil {
-		renderError(rw, err)
-		return
-	}
-
-	response := struct {
-		Count int64           `json:"count"`
-		Jobs  []*work.DeadJob `json:"jobs"`
-	}{Count: count, Jobs: jobs}
-
-	render(rw, response, err)
-}
-
-func (c *context) deleteDeadJob(rw web.ResponseWriter, r *web.Request) {
-	diedAt, err := strconv.ParseInt(r.PathParams["died_at"], 10, 64)
-	if err != nil {
-		renderError(rw, err)
-		return
-	}
-
-	err = c.client.DeleteDeadJob(diedAt, r.PathParams["job_id"])
-
-	render(rw, map[string]string{"status": "ok"}, err)
-}
-
-func (c *context) retryDeadJob(rw web.ResponseWriter, r *web.Request) {
-	diedAt, err := strconv.ParseInt(r.PathParams["died_at"], 10, 64)
-	if err != nil {
-		renderError(rw, err)
-		return
-	}
-
-	err = c.client.RetryDeadJob(diedAt, r.PathParams["job_id"])
-
-	render(rw, map[string]string{"status": "ok"}, err)
-}
-
-func (c *context) deleteAllDeadJobs(rw web.ResponseWriter, r *web.Request) {
-	err := c.client.DeleteAllDeadJobs()
-	render(rw, map[string]string{"status": "ok"}, err)
-}
-
-func (c *context) retryAllDeadJobs(rw web.ResponseWriter, r *web.Request) {
-	err := c.client.RetryAllDeadJobs()
-	render(rw, map[string]string{"status": "ok"}, err)
 }
 
 func render(rw web.ResponseWriter, jsonable interface{}, err error) {
