+++ webui/webui_test.go
@@ -34,29 +34,18 @@ func TestWebUIPing(t *testing.T) {
 	cleanKeyspace(ns, pool)
 
 	t.Run("with inbuilt server", func(t *testing.T) {
-		t.Run("vanilla server", func(t *testing.T) {
-			s := NewServer(ns, pool, ":6666")
+		s := NewServer(ns, pool, ":6666")
 
-			recorder := httptest.NewRecorder()
-			request, _ := http.NewRequest("GET", "/ping", nil)
-			s.router.ServeHTTP(recorder, request)
-			assert.Equal(t, 200, recorder.Code)
-		})
-
-		t.Run("with prefix", func(t *testing.T) {
-			s := NewServer(ns, pool, ":6666", WithPrefix("/api"))
-
-			recorder := httptest.NewRecorder()
-			request, _ := http.NewRequest("GET", "/api/ping", nil)
-			s.router.ServeHTTP(recorder, request)
-			assert.Equal(t, 200, recorder.Code)
-		})
+		recorder := httptest.NewRecorder()
+		request, _ := http.NewRequest("GET", "/ping", nil)
+		s.router.ServeHTTP(recorder, request)
+		assert.Equal(t, 200, recorder.Code)
 	})
 
 	t.Run("with external server", func(t *testing.T) {
-		t.Run("vanilla server", func(t *testing.T) {
-			unstartedWorkerUI := NewServer(ns, pool, ":6666")
-			testServer := httptest.NewServer(unstartedWorkerUI.Router())
+		t.Run("no prefix", func(t *testing.T) {
+			router := NewRouter(work.NewClient(ns, pool), RouterOptions{})
+			testServer := httptest.NewServer(router)
 
 			request, err := http.NewRequest("GET", fmt.Sprintf("%s/ping", testServer.URL), nil)
 			require.NoError(t, err)
@@ -68,8 +57,8 @@ func TestWebUIPing(t *testing.T) {
 		})
 
 		t.Run("with prefix", func(t *testing.T) {
-			unstartedWorkerUI := NewServer(ns, pool, ":6666", WithPrefix("/api"))
-			testServer := httptest.NewServer(unstartedWorkerUI.Router())
+			router := NewRouter(work.NewClient(ns, pool), RouterOptions{PathPrefix: "/api"})
+			testServer := httptest.NewServer(router)
 
 			request, err := http.NewRequest("GET", fmt.Sprintf("%s/api/ping", testServer.URL), nil)
 			require.NoError(t, err)
@@ -524,18 +513,19 @@ func TestWebUIAssets(t *testing.T) {
 	t.Run("router has prefix", func(t *testing.T) {
 		pool := newTestPool(t)
 		ns := "testwork"
-		s := NewServer(ns, pool, ":6666", WithPrefix("/prefix"))
+
+		router := NewRouter(work.NewClient(ns, pool), RouterOptions{PathPrefix: "/prefix"})
 
 		recorder := httptest.NewRecorder()
 		request, _ := http.NewRequest("GET", "/prefix", nil)
-		s.router.ServeHTTP(recorder, request)
+		router.ServeHTTP(recorder, request)
 		body := string(recorder.Body.Bytes())
 		assert.Regexp(t, "html", body)
 		assert.Regexp(t, `src="/prefix/work.js"`, body)
 
 		recorder = httptest.NewRecorder()
 		request, _ = http.NewRequest("GET", "/prefix/work.js", nil)
-		s.router.ServeHTTP(recorder, request)
+		router.ServeHTTP(recorder, request)
 	})
 }
